+++ src/pages/Report.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import DataSdk from '../SDK/DataSdk';
 import ReportChart from '../components/ReportChart';
 import ReportInfo from "../components/ReportIntro";
+import LinkButton from "../components/LinkButton";
 import { useLocation } from 'react-router-dom'
 
 const dataSdk = new DataSdk();
@@ -50,25 +51,6 @@ const Report = (props) => {
         console.log(`Loading report page for ${sraAccession}.`);
         fetchEntrezData();
     }, [sraAccession]);
-
-    const LinkButton = (props) => {
-        let aAttrs = {};
-        if (props.newTab) {
-            aAttrs = {
-                target: "_blank",
-                rel: "noopener noreferrer"
-            };
-        }
-        return (
-            <button className="bg-gray-300 hover:bg-gray-500 mx-2 py-2 px-4 rounded inline-flex items-center">
-                <a className="text-blue-500" {...aAttrs}
-                    href={props.link} download={props.download}>
-                    {props.text}
-                    {props.icon}
-                </a>
-            </button>
-        )
-    }
 
     let downloadIcon = (<svg className="inline fill-current w-4 h-4 ml-1 mb-1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"><path d="M13 8V2H7v6H2l8 8 8-8h-5zM0 18h20v2H0v-2z" /></svg>)
     let externalLinkIcon = (<svg className="inline fill-current w-4 h-4 ml-1 mb-1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M19,21H5c-1.1,0-2-0.9-2-2V5c0-1.1,0.9-2,2-2h7v2H5v14h14v-7h2v7C21,20.1,20.1,21,19,21z" /><path d="M21 10L19 10 19 5 14 5 14 3 21 3z" /><path d="M6.7 8.5H22.3V10.5H6.7z" transform="rotate(-45.001 14.5 9.5)" /></svg>)
